Major refactoring of the test classes
Large changes, a lot changed. I was not satisfied with the execution
speed of the tests, mainly cause of the many different application
contexts that started. I split the tests into it (integration tests) and
ut (unit tests) and put pretty everything into the ut.
The it are just used to test the controller, the interface we expose to
the outside world. This way the time needed to run the tests dropped and
went down to an acceptable amount.
But this all comes to a price: Any other RefreshToken implementation has
to change his tests to the new layout, but this only requires a couple
of small changes.